# backend/docs/swagger.yaml
definitions:
...
  models.Recipe:
    properties:
      cook_time:
        type: integer
      recipe_description:
        type: string
      recipe_id:
        type: integer
      recipe_name:
        type: string
    type: object
...
  models.RecipeRequest:
    properties:
      cook_time:
        type: integer
      recipe_description:
        type: string
      recipe_name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /recipes:
    get:
      consumes:
      - application/json
      description: Get a list of all recipes from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Recipe'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all recipes
      tags:
      - recipes
    post:
      consumes:
      - application/json
      description: Add a new recipe to the database
      parameters:
      - description: Add recipe
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/models.RecipeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new recipe
      tags:
      - recipes
  /recipes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a recipe from the database by ID
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Recipe deleted
        "400":
          description: Invalid recipe ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Recipe not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a recipe by ID
      tags:
      - recipes
    get:
      consumes:
      - application/json
      description: Get a single recipe from the database by ID
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a recipe by ID
      tags:
      - recipes
    put:
      consumes:
      - application/json
      description: Update an existing recipe by ID
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update recipe
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/models.Recipe'
      produces:
      - application/json
      responses:
        "200":
          description: Recipe updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid recipe ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Recipe not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update a recipe by ID
      tags:
      - recipes
swagger: "2.0"